Move deferred to utils; Remove TODOs
